requirements.txt
Django==3.2.7
django-haystack==3.1.1
mysqlclient==2.1.0
Pillow==8.4.0
python-dotenv==0.19.2
Whoosh==2.7.4
